User: Kay          Date:  3.07.08   Time: 23:44 (vss version: 3, vss user: Kay)

+++ SuRun/LSALogon.cpp
@@ -19,7 +19,7 @@
 
 void InitString(PLSA_STRING LsaString,LPSTR String)
 {
-  DWORD StringLength;
+  size_t StringLength;
   if (String == NULL)
   {
     LsaString->Buffer = NULL;
